template/requirements.txt
asgiref
autopep8
django==2.2.10
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
django-cors-headers
pyserial==3.4
tornado
djangorestframework